+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+## Fixed
+- Parsing `artisan route:list` when Docker also prints to the console ([\#92](https://github.com/NiclasvanEyk/jetbrains-laravel-make-integration/issues/92))
+
 ## 3.4.0 - 2022-12-06
 
 ### Added